Fix pika upstream
