+++ env.sample
@@ -1,4 +1,4 @@
 REDIS_HOST=redis
 REDIS_PORT=6379
 TDS_URL=http://data-service-api:8000
-PYCIEMSS_OUTPUT_FILEPATH=result.csv
+RABBITMQ_HOST=rabbitmq.pyciemss
