+++ service/models.py
@@ -186,4 +186,3 @@ class JobResponse(BaseModel):
 
 class StatusSimulationIdGetResponse(BaseModel):
     status: Optional[Status] = None
-    progress: Optional[float]
