+++ service/settings.py
@@ -10,9 +10,10 @@ class Settings(BaseSettings):
     """
     pyciemss-service configuration
     """
+    TDS_URL: str = "http://data-service-api:8000"
     REDIS_HOST: str = "redis"
     REDIS_PORT: int = 6379
-    TDS_URL: str = "http://data-service-api:8000"
+    RABBITMQ_HOST: str = "rabbitmq.pyciemss"
 
 
 settings = Settings()
